fix: added correct logic for sidebar state
