+++ src/components/shared/MuiLink/index.tsx
@@ -2,11 +2,12 @@ import React from 'react';
 import { Link as MaterialLink } from '@mui/material';
 import { Link as RouterLink } from 'react-router-dom';
 
-import { LinkProps } from './types';
+import { ILinkProps } from './types';
 
-function MuiLink({ children, to = '/', ...props }: LinkProps): JSX.Element {
+function MuiLink({ children, to = '/', underline = 'always', ...props }: ILinkProps): JSX.Element {
+  console.log({ ...props });
   return (
-    <MaterialLink component={RouterLink} to={to} {...props}>
+    <MaterialLink component={RouterLink} to={to} underline={underline} {...props}>
       {children}
     </MaterialLink>
   );
