--- src/pages/Dashboard/Header/HeaderDrawer/ListItem/types.ts
@@ -1,6 +1,0 @@
-import { RoutesData } from '../types';
-
-export interface ListItemProps extends RoutesData {
-  isListItemOpen: boolean;
-  setListItemOpen: (value: boolean) => void;
-}
