+++ src/pages/Dashboard/Header/HeaderDrawer/SideBarList/index.ts
@@ -1,0 +1,1 @@
+export { SideBarList } from './component';
